golang: extract top-level race-detector config for tests
It's particularly important to enable the race-detector in CI tests,
because it's able to catch and flag correctness issues which may not
show up for folks running vanilla `go test` locally.

It also seems likely that teams will tweak the default test options
(e.g., to increase the timeout) and accidentally remove the `-race` flag
without understanding the implications.

This extracts a new top-level config `race-detector` for tests which
controls the `-race` flag to call attention to its presence and make it
harder to accidentally remove.