Add check-modules (true) and check-vendor (false)
check-modules runs go mod {tidy,verify} and checks to see if git is
still clean. If it's not then it's an error.

check-vendor runs go mod vendor and checks to see if git is still clean.
If it's not then it's an error.

This required extracting the cache logic from the go command into the
job itself.

[W-6710181](https://gus.lightning.force.com/lightning/r/ADM_Work__c/a07B00000076rAOIAY/view)