Updating Snyk orb to fix download Snyk CLI (#23)
The old version of Snyk orb used cUrl to download the Snyk CLI client,
but something changed on GitHub that required adding the follow
redirects to that cUrl command. Checking the code of the latest version
of Snyk orb, that cUrl was changed to another URL. That will fix the
current issue and also keep us up to date.